@@ -25,13 +25,13 @@
     ]
   },
   "devDependencies": {
-    "@fluentui/azure-themes": "^8.1.126",
-    "@fluentui/react": "^8.36.3",
+    "@fluentui/azure-themes": "^8.2.52",
+    "@fluentui/react": "^8.77.2",
     "@microsoft/applicationinsights-web": "^2.7.0",
     "@playwright/test": "^1.15.2",
     "playwright-expect": "^0.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1"
   }
 }
