@@ -21,7 +21,7 @@ export function getError(array, field) {
 export const adv_stackstyle = { root: { border: "1px solid", margin: "10px 0", padding: "15px" } }
 
 
-export function CodeBlock({deploycmd, testId, lang, filename}) {
+export function CodeBlock({deploycmd, testId, lang, filename, error}) {
     const [ copied, setCopied ] = useState(false)
 
     function copyIt() {
@@ -36,21 +36,28 @@ export function CodeBlock({deploycmd, testId, lang, filename}) {
     }
 
     return [
-        <div key="code-header" className="codeHeader" >
+        <div key="code-header" className="codeHeader" style={{...(error && {borderColor: 'darkred'})}}>
             <span className="language">{lang}</span>
+            { error && <div  className="error">{error}</div> }
             <CommandBarButton
-              //disabled={filename.length()>0}
+              disabled={error}
               className="action position-relative"
               iconProps={{ iconName: copied? 'Completed' : 'Save'}}
-              styles={{icon: {color: '#171717'}}}
-              text="Save"
+              //styles={{icon: {color: '#171717'}}}
+              text={!error ? "Save" : ""}
               primaryActionButtonProps={{download: filename}}
               onClick={downloadIt}/>
-            <CommandBarButton disabled={copied} className="action position-relative" iconProps={{ iconName: copied? 'Completed' : 'Copy'}} styles={{icon: {color: '#171717'}}} text="Copy" onClick={copyIt}/>
+            <CommandBarButton
+                    disabled={copied || error}
+                    className="action position-relative"
+                    iconProps={{ iconName: copied? 'Completed' : 'Copy'}}
+                    styles={{icon: {color: '#171717'}}}
+                    text={!error ? "Copy" : ""}
+                    onClick={copyIt}/>
         </div>,
 
-        <pre key="code-pre" className="has-inner-focus">
-            <code className={"lang-" + lang}><span data-lang={lang} data-testid={testId || 'none'}>{deploycmd}</span></code>
+        <pre key="code-pre" className="has-inner-focus"  style={{...(error && {borderColor: 'darkred'})}}>
+            <code className={"lang-" + lang}><span data-lang={lang} data-testid={testId || 'none'} style={{...(error && {color: 'lightgrey'})}}>{deploycmd}</span></code>
         </pre>
     ]
 }
